chore: refactor DBSearchPage
