+++ packages/app/src/components/DBSearchPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DBSearchPage';
+export { DBSearchPage } from './DBSearchPage';
